Fix realpath error on macOS (#7)
* Fix realpath error on macOS

Run `grealpath` instead of `realpath` on macOS to explicitly use the
GNU version provided by the `coreutils` formula. This is needed because
the built-in macOS version of realpath does not support the required
flags.

* Implement realpath fallback on macOS

If GNU realpath is not available as `grealpath`, fall back to
`realpath` on macOS and assume that this is what the user has GNU
realpath installed as.

* More

* Update runner

---------

Co-authored-by: Radon Rosborough <radon@intuitiveexplanations.com>